requirements.txt
flask
gunicorn
krakenex
python-dotenv


